src/test/java/CrackingCodingInterview/Trees/MaximumInBinaryTreeTest.java
package CrackingCodingInterview.Trees;

import org.testng.Assert;
import org.testng.annotations.Test;

import static org.testng.Assert.*;

/**
 * Created by root on 1/9/16.
 */
public class MaximumInBinaryTreeTest {
    MaximumInBinaryTree m = new MaximumInBinaryTree();

    @Test
    public void testFindMaxInBinaryTree() throws Exception {
        BinaryTreeNode root = new BinaryTreeNode(1);
        BinaryTreeNode left = new BinaryTreeNode(2);
        BinaryTreeNode right = new BinaryTreeNode(3);
        BinaryTreeNode left1 = new BinaryTreeNode(4);
        BinaryTreeNode left2 = new BinaryTreeNode(5);
        BinaryTreeNode right1 = new BinaryTreeNode(6);

        root.setLeft(left);
        root.setRight(right);
        left.setLeft(left1);
        left.setRight(left2);
        right.setLeft(right1);

        int max = m.findMaxInBinaryTree(root);
        int data = (Integer) right1.getData();
        Assert.assertEquals(max , data);
    }

    @Test
    public void testFindMaxInBinaryTree2() throws Exception {
        BinaryTreeNode root = new BinaryTreeNode(1);
        BinaryTreeNode left = new BinaryTreeNode(2);
        BinaryTreeNode right = new BinaryTreeNode(3);
        BinaryTreeNode left1 = new BinaryTreeNode(4);
        BinaryTreeNode left2 = new BinaryTreeNode(5);
        BinaryTreeNode right1 = new BinaryTreeNode(6);

        root.setLeft(left);
        root.setRight(right);
        left.setLeft(left1);
        left.setRight(left2);
        right.setLeft(right1);

        int max = m.findMaxInBinaryTreeWithoutRecursion(root);
        int data = (Integer) right1.getData();
        Assert.assertEquals(max , data);
    }
}